Updated the code for more readability and  hiding the chat processing and modified "is_ollama_installed" in ollama_installer.py
